[Fleet] Display warning when trying to upgrade agent to version > max fleet server version (#178079)
Closes [173727](https://github.com/elastic/kibana/issues/173727)
## Summary

Adding a warning to the agent upgrade modal: when the selected version
is > than the greatest installed fleet server agent, a warning is
presented to the user.

This happens both for a single agent upgrade and for a bulk upgrade. The
submit button is only disabled in the case of a single agent upgrade to
avoid blocking the user from doing a bulk upgrade. In that case the
endpoint will respond with the same error.

Also adding a link to agent upgrade docs and opened a request to improve
the docs: https://github.com/elastic/ingest-docs/issues/966

![Screenshot 2024-03-11 at 10 39
49](https://github.com/elastic/kibana/assets/16084106/5f3765b6-af25-4fa1-99fd-993d561d0f3b)

![Screenshot 2024-03-08 at 17 17
26](https://github.com/elastic/kibana/assets/16084106/1d0cfe26-8d6e-4937-86da-a53072069450)

Warning for "not upgradeable agent":
![Screenshot 2024-03-11 at 10 50
10](https://github.com/elastic/kibana/assets/16084106/249c69c1-0e38-4c73-b2bf-8122a562da49)




### Checklist

- [ ]
[Documentation](https://www.elastic.co/guide/en/kibana/master/development-documentation.html)
was added for features that require explanation or tutorials
- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [ ] Any UI touched in this PR does not create any new axe failures
(run axe in browser:
[FF](https://addons.mozilla.org/en-US/firefox/addon/axe-devtools/),
[Chrome](https://chrome.google.com/webstore/detail/axe-web-accessibility-tes/lhdoppojpmngadmnindnejefpokejbdd?hl=en-US))

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Julia Bardi <90178898+juliaElastic@users.noreply.github.com>