[kbn-journey] keep ES data b/w test & warmup runs (#178388)
## Summary

When we run journey to collect EBT/AMP metrics as part of performace
pipeline, we use the `scripts/run_performance.js` script, that does the
following:

1. Start Elasticsearch
2. Start Kibana with APM on / EBT off, run the journey scenario, stop
Kibana
3. Start Kibana again with both APM/EBT on, run journey scenario, stop
Kibana
4. Stop Elasticsearch

Step 3 is where the monitoring metrics are collected. Step 2 is only to
"warmup" ES, experimentally proved that metrics are more consistent when
we run the journey 2 times.
On the other hand journey lifecycle has `before` & `after` hooks that
handle test data adding/removing and it doesn't help to unload ES data
after WARMUP phase and load again before TEST one.
Still journey can and is run as a functional test, so we need to keep
"cleanup" for when it is run locally e.g. to debug the test flow not the
metrics collection.

Relying on TEST_PERFORMANCE_PHASE env var that is set only during
performance run, this PR makes a change to avoid unloading ES data after
warmup and ingest data before test only if index was deleted (*should
never happen, but for validation purpose). This way we ingest data to
Elasticsearch only one time.